chore: Update Dockerfile to use latest node image and install AWS CLI
This commit updates the Dockerfile to use the latest node image (version 22.2.0) as the base image. It also installs the AWS CLI on top of it. This change was made to resolve dependency issues between the Amazon Linux and the latest versions of Node, specifically related to the groff package. The working directory in the container is set to /usr/src/app, and additional dependencies (python3, python3-pip, groff-base) are installed. The AWS CLI is then downloaded and installed using the provided script.