chore: Update Dockerfile to include session-manager-plugin and CLI services
This commit updates the Dockerfile to include the installation of the session-manager-plugin and CLI services. It also copies the ./cli_services directory into the container and makes all .sh files in the services directory and its subdirectories executable. Additionally, it installs the CLI pager and the jq tool for parsing JSON in shell scripts. The working directory in the container is set to /usr/src/app/cli_services.

```