feat: Add script to interactively SSH into EC2 instances
This commit adds a new script, ssh_connect.sh, which allows users to SSH into EC2 instances interactively. The script retrieves a list of running instances in the current region and prompts the user to select an instance from the list. Once an instance is selected, the script uses AWS SSM to start a session with the selected instance. This feature improves the user experience by providing a convenient way to access EC2 instances without manually running AWS CLI commands.